+++ Source/Workers/decodeGoogleEarthEnterprisePacket.js
@@ -20,7 +20,8 @@ define([
 
     var Types = {
         METADATA : 0,
-        TERRAIN : 1
+        TERRAIN : 1,
+        DBROOT : 2
     };
 
     Types.fromString = function(s) {
@@ -28,13 +29,15 @@ define([
             return Types.METADATA;
         } else if (s === 'Terrain') {
             return Types.TERRAIN;
+        } else if (s === 'DbRoot') {
+            return Types.DBROOT;
         }
     };
 
-    function decodeGoogleEarthEnterpriseTerrainPacket(parameters, transferableObjects) {
+    function decodeGoogleEarthEnterprisePacket(parameters, transferableObjects) {
         var type = Types.fromString(parameters.type);
         var buffer = parameters.buffer;
-        decodeGoogleEarthEnterpriseData(buffer);
+        decodeGoogleEarthEnterpriseData(parameters.key, buffer);
 
         var uncompressedTerrain = uncompressPacket(buffer);
         buffer = uncompressedTerrain.buffer;
@@ -45,6 +48,11 @@ define([
                 return processMetadata(buffer, length, parameters.quadKey);
             case Types.TERRAIN:
                 return processTerrain(buffer, length, transferableObjects);
+            case Types.DBROOT:
+                transferableObjects.push(buffer);
+                return {
+                    buffer : buffer
+                };
         }
 
     }
@@ -131,13 +139,13 @@ define([
 
             offset += 8; // Ignore image neighbors for now
 
-            // Data providers aren't used
-            ++offset; // Image provider
-            ++offset; // Terrain provider
+            // Data providers
+            var imageProvider = dv.getUint8(offset++);
+            var terrainProvider = dv.getUint8(offset++);
             offset += sizeOfUint16; // 4 byte align
 
             instances.push(new GoogleEarthEnterpriseTileInformation(bitfield, cnodeVersion,
-                imageVersion, terrainVersion));
+                imageVersion, terrainVersion, imageProvider, terrainProvider));
         }
 
         var tileInfo = [];
@@ -243,5 +251,5 @@ define([
         return uncompressedPacket;
     }
 
-    return createTaskProcessorWorker(decodeGoogleEarthEnterpriseTerrainPacket);
+    return createTaskProcessorWorker(decodeGoogleEarthEnterprisePacket);
 });